refactor: move dropdown to components
