+++ src/atoms/index.ts
@@ -3,4 +3,3 @@ export { default as Tooltip } from './tooltip/Tooltip';
 export * from './icons';
 export * from './button/types';
 export * from './tooltip/types';
-export { default as DropdownMenu } from './dropdownMenu/DropdownMenu';
