+++ src/components/dropdown/DropdownMenu.mdx
@@ -4,8 +4,8 @@ menu: Components
 ---
 
 import { Playground, Props } from 'docz';
-import DropdownMenu from '@atoms/dropdownMenu/DropdownMenu';
-import DropdownElement from '@ions/dropdown/element/DropdownElement';
+import DropdownMenu from '@components/dropdown/DropdownMenu';
+import DropdownElement from '@components/dropdown/DropdownElement';
 import Button from '@atoms/button/Button';
 import PhoneIcon from '@atoms/icons/PhoneIcon';
