+++ src/components/index.ts
@@ -1,0 +1,2 @@
+export { default as DropdownMenu } from './dropdown/DropdownMenu';
+export { default as DropDownElement } from './dropdown/DropDownElement';
