+++ src/index.ts
@@ -1,3 +1,3 @@
 export * from './atoms';
 export * from './molecules';
-export { DropdownElement } from './ions';
+export * from './components';
