+++ src/ions/index.ts
@@ -1,6 +1,3 @@
 export { default as Icon } from '@ions/icon/base/Icon';
 export { default as SvgWrapper } from '@ions/icon/wrapper/SvgWrapper';
 export { default as BaseButton } from '@ions/button/base/BaseButton';
-export {
-  default as DropdownElement
-} from '@ions/dropdown/element/DropdownElement';
